+++ mirage/factories/identifier.ts
@@ -8,3 +8,9 @@ export default Factory.extend<Identifier>({
         return faker.fake('10.5555/{{company.bsNoun}}');
     },
 });
+
+declare module 'ember-cli-mirage/types/registries/schema' {
+    export default interface MirageSchemaRegistry {
+        identifierss: Identifier;
+    } // eslint-disable-line semi
+}