+++ mirage/factories/registration-schema.ts
@@ -2,8 +2,7 @@ import { Factory, faker } from 'ember-cli-mirage';
 
 import RegistrationSchema from 'ember-osf-web/models/registration-schema';
 
-export interface MirageRegistrationSchema extends
-    Pick<RegistrationSchema, Exclude<keyof RegistrationSchema, 'schema'>> {
+export interface MirageRegistrationSchema extends RegistrationSchema {
     schemaNoConflict?: RegistrationSchema['schema'];
 }
 
@@ -32,5 +31,11 @@ export default Factory.extend<MirageRegistrationSchema>({
 declare module 'ember-cli-mirage/types/registries/model' {
     export default interface MirageModelRegistry {
         'registration-schema': MirageRegistrationSchema;
+    } // eslint-disable-line semi
+}
+
+declare module 'ember-cli-mirage/types/registries/schema' {
+    export default interface MirageSchemaRegistry {
+        registrationSchemas: MirageRegistrationSchema;
     } // eslint-disable-line semi
 }