+++ mirage/views/developer-app.ts
@@ -1,18 +1,16 @@
 import { faker, HandlerContext, Request, Schema } from 'ember-cli-mirage';
 
-import DeveloperApp from 'ember-osf-web/models/developer-app';
-
 export function createDeveloperApp(this: HandlerContext, schema: Schema) {
     const attrs = {
-        ...this.normalizedRequestAttrs('developerApp'),
+        ...this.normalizedRequestAttrs('developer-app'),
         clientId: faker.internet.ip(),
         clientSecret: faker.random.uuid(),
     };
-    return schema.developerApps.create<DeveloperApp>(attrs);
+    return schema.developerApps.create(attrs);
 }
 
 export function resetClientSecret(this: HandlerContext, schema: Schema, request: Request) {
-    const developerApp = schema.developerApps.find<DeveloperApp>(request.params.id);
+    const developerApp = schema.developerApps.find(request.params.id);
     developerApp.update('clientSecret', faker.random.uuid());
     return this.serialize(developerApp);
 }