+++ mirage/views/fork.ts
@@ -1,13 +1,10 @@
-import { HandlerContext, ModelInstance, Schema } from 'ember-cli-mirage';
-
-import Node from 'ember-osf-web/models/node';
-import Registration from 'ember-osf-web/models/registration';
+import { HandlerContext, Schema } from 'ember-cli-mirage';
 
 export function createFork(this: HandlerContext, schema: Schema) {
     const attrs = this.normalizedRequestAttrs('node');
     const nodeId = attrs.id;
     delete attrs.id;
-    const node = schema.nodes.find(nodeId) as ModelInstance<Node>;
+    const node = schema.nodes.find(nodeId);
     const fork = schema.nodes.create({
         forkedFrom: node,
         category: node.category,
@@ -18,7 +15,7 @@ export function createFork(this: HandlerContext, schema: Schema) {
     });
 
     const userId = schema.roots.first().currentUserId;
-    const currentUser = schema.users.find(userId);
+    const currentUser = schema.users.find(userId!);
     node.contributors.models.forEach(contributor => {
         schema.contributors.create({ node: fork, users: contributor.users });
     });
@@ -31,8 +28,8 @@ export function createRegistrationFork(this: HandlerContext, schema: Schema) {
     const attrs = this.normalizedRequestAttrs('node');
     const registrationId = attrs.id;
     delete attrs.id;
-    const registration = schema.nodes.find(registrationId) as ModelInstance<Registration>;
-    const fork = schema.registrations.create({
+    const registration = schema.registrations.find(registrationId);
+    const fork = schema.nodes.create({
         forkedFrom: registration,
         category: registration.category,
         fork: true,
@@ -42,7 +39,7 @@ export function createRegistrationFork(this: HandlerContext, schema: Schema) {
     });
 
     const userId = schema.roots.first().currentUserId;
-    const currentUser = schema.users.find(userId);
+    const currentUser = schema.users.find(userId!);
     registration.contributors.models.forEach(contributor => {
         schema.contributors.create({ node: fork, users: contributor.users });
     });