+++ mirage/views/guid.ts
@@ -1,13 +1,11 @@
 import { HandlerContext, Request, Response, Schema } from 'ember-cli-mirage';
 
-import Guid from 'ember-osf-web/models/guid';
-
 import { queryParamIsTruthy } from './utils';
 
 export function guidDetail(this: HandlerContext, schema: Schema, request: Request) {
     const { id } = request.params;
     const { resolve } = request.queryParams;
-    const guid = schema.guids.find<Guid>(id);
+    const guid = schema.guids.find(id);
 
     if (!guid) {
         return new Response(404, {}, {