+++ mirage/views/registration.ts
@@ -1,15 +1,12 @@
 import { HandlerContext, Request, Response, Schema } from 'ember-cli-mirage';
 
-import Registration from 'ember-osf-web/models/registration';
-import RegistrationSchema from 'ember-osf-web/models/registration-schema';
-
 import { process } from './utils';
 
 export function forkRegistration(this: HandlerContext, schema: Schema) {
-    const attrs = this.normalizedRequestAttrs();
-    const forkedFrom = schema.registrations.find<Registration>(attrs.id);
-    const registrationSchema = schema.registrationSchemas.find<RegistrationSchema>('prereg_challenge');
-    const newFork = schema.registrations.create<Registration>({
+    const attrs = this.normalizedRequestAttrs('registration');
+    const forkedFrom = schema.registrations.find(attrs.id);
+    const registrationSchema = schema.registrationSchemas.find('prereg_challenge');
+    const newFork = schema.registrations.create({
         forkedDate: new Date(),
         fork: true,
         forkedFrom,
@@ -21,7 +18,7 @@ export function forkRegistration(this: HandlerContext, schema: Schema) {
 
 export function registrationDetail(this: HandlerContext, schema: Schema, request: Request) {
     const { id } = request.params;
-    const registration = schema.registrations.find<Registration>(id);
+    const registration = schema.registrations.find(id);
 
     if (registration.embargoed && !registration.currentUserPermissions.length) {
         return new Response(404, {}, {